Done sorting, delete, Todo: content view, edit

+++ frontend/src/components/mails/MailItem.js
@@ -6,18 +6,33 @@ import { connect } from 'react-redux'
 import M from 'materialize-css/dist/js/materialize.min.js'
 
 const MailItem = ({ mail, deleteMail, setCurrent }) => {
-    
     const onDelete = () => {
-        deleteMail(mail.id);
-        M.toast({ html: 'Log Deleted'})
+        if(window.confirm('Na pewno chcesz usunąć ten E-mail?')) {
+            deleteMail(mail.id);
+            M.toast({ html: 'E-mail został usunięty'});
+        }
+    }
+
+    const shortContent = (content) => { 
+        if(content.length > 30){
+            let shortedContent = (content.substring(0,30))
+            return shortedContent.substring(0, shortedContent.lastIndexOf(' ')) + ' ...'
+        } else {
+            return content
+        }
+
+        
     }
 
     return (       
         <tr>
             <td>{mail.title}</td>
-            <td>{mail.content}</td>
+            <td>{shortContent(mail.content)}</td>
             <td>{mail.date}</td>
 
+            <td className="material-icons">
+                visibility
+            </td>
             <td className="material-icons">
                 edit
             </td>
